update info of the peers when trying to add an already present peer + rename api-server folders
